task-1894: the high and medium priority findings from the Inillucent code review
H1 through H3 and M1 through M8 from the review Codex Sol produced, each
implemented and each with a test that fails when the fix is reverted.

- H1 `--root` is enforced against the file system rather than against path text.
  `inillucent-vfs/src/confine.rs` is the one authorization service: it resolves
  every component, so a junction or a symbolic link cannot carry a path out of
  the root, and `..` pops the resolved path rather than the named one. `open`,
  `delete`, `access` and `temp_path` go through it, and `ATTACH DATABASE` and
  `VACUUM INTO` refuse by name before the engine opens anything. Twelve end to
  end cases drive the real `inillucent` and `inillucent-mcp` binaries
- H2 the C API's handles are safe in every destruction order. A connection is an
  `Rc<ConnState>` that statements and transactions share, so freeing a
  connection with live children is now defined behaviour rather than a dangling
  pointer, and the database's connection count falls when the last child goes.
  The header's ownership section says the new rule
- H3 PostgreSQL and MySQL migration runs over verified TLS. SChannel on Windows
  with an exclusive root chain engine, OpenSSL through `dlopen` on Unix, a TLS
  1.2 floor and a required hostname check on both. Plaintext to a network
  address needs `--insecure-plaintext`; loopback is unchanged
- M1 `rust-toolchain.toml` pins 1.95.0, `clippy.toml` carries the thresholds,
  and `.github/workflows/ci.yml` plus `tools/validate.ps1` and `tools/validate.sh`
  run the same checks on both platforms
- M2 a C conformance program exercises every handle type, every null, and four
  destruction orders, built and run twice - plain and under AddressSanitizer
- M3 the shadow store is a `Pager` again, so `inillucent-vm` no longer reaches
  into the retired storage and transaction crates. `policy.rs` has a ratchet
  that fails if a new crate does
- M4 `ImportedDatabase::import_into` was 493 lines doing seven things; it is now
  seven named phases in `inillucent-engine/src/import.rs`, and a size ratchet
  records fourteen modules so none of them grows back
- M5 MCP requests carry a budget: rows, bytes, time, memory and cancellation,
  read at `Collect::push` and at every leaf of the paged executor, reported as
  structured errors
- M6 release creation verifies the source state, the version, the tag, the CI
  result, the packages, the installations, the checksums and the provenance, and
  runs command line, MCP, reopen, backup, restore and C API smoke checks
- M7 `docs/relational-architecture.md` describes the engine as it is now, and the
  stale Nikaya recovery report is corrected
- M8 an incremental vector insert no longer rebuilds the graph. Automatic
  compaction folds the delta log into the published generation - one graph
  insert per delta entry rather than one per row in the table - and the
  single-pass build is what the `compact` and `rebuild` commands run. Measured
  by `inillucent-foldgate` at 40,000 documents: 35,579 graph inserts against
  255,989, a worst single publish of 4,447 against 35,579, recall at ten of
  0.704 against 0.637, and 180.9 MiB peak against 196.4

Performance check: four consecutive medium gate runs, 30 paired rounds each,
medians of the two middle runs, against the figures docs/performance.md
publishes. Headline 4.46x weighted against 4.26x. Nine of the ten families are
at or above their published ratio and none is below the contract floor;
read.analytical reads 6.82x against 6.94x, inside the 6.79 to 6.85 spread of the
four runs. Peak resident memory 42.20 MiB against 42.61. Processor time 445 ms
against 422, inside a 406 to 469 spread and still under the contract bar on
every run. Run files are in _agent_output/task-1894-perfcheck.

TDD: /notebook?noteId=7d0bf131-cbcb-4396-a0fb-92b7cc59db05

+++ crates/inillucent-scalar/src/regexp.rs
@@ -256,7 +256,7 @@ impl Compiler<'_> {
         let mut value = 0u32;
         for offset in 1..=count {
             let digit = self.text.get(self.at + offset).copied()?;
-            value = value * 16 + u32::from(char::from(digit).to_digit(16)?);
+            value = value * 16 + char::from(digit).to_digit(16)?;
         }
         Some(value)
     }
